+++ back/app/controllers/application_controller.rb
@@ -4,7 +4,7 @@ class ApplicationController < ActionController::API
   include Knock::Authenticable
   include Pundit
 
-  before_action :authenticate_user, if: :secure_controller?
+  before_action :authenticate_user
   after_action :verify_authorized, except: :index
   after_action :verify_policy_scoped, only: :index
 
@@ -19,11 +19,6 @@ class ApplicationController < ActionController::API
 
   rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
-  # all controllers are secured by default
-  def secure_controller?
-    true
-  end
-
   def send_success(data = nil, status = 200)
     render json: data, status: status
   end
@@ -34,9 +29,9 @@ class ApplicationController < ActionController::API
 
   def send_not_found(error = nil)
     if error.nil?
-      head 404, 'content_type' => 'text/plain'
+      head :not_found, 'content_type' => 'text/plain'
     else
-      render json: error, status: 404
+      render json: error, status: :not_found
     end
   end
 
@@ -98,5 +93,10 @@ class ApplicationController < ActionController::API
     pageparams['page'] ||= {}
     pageparams['page']['number'] = number
     "#{url}?#{pageparams.to_param}"
+  end
+
+  def paginate(collection)
+    collection.page(params.dig(:page, :number))
+              .per(params.dig(:page, :size))
   end
 end