+++ back/spec/models/static_page_spec.rb
@@ -1,0 +1,9 @@
+require 'rails_helper'
+
+RSpec.describe StaticPage, type: :model do
+  describe 'Default factory' do
+    it 'is valid' do
+      expect(build(:static_page)).to be_valid
+    end
+  end
+end