+++ front/app/modules/commercial/customizable_navbar/admin/components/ModuleActive.tsx
@@ -1,0 +1,14 @@
+import { useEffect } from 'react';
+
+interface Props {
+  onMount: () => void;
+}
+
+export default ({ onMount }: Props) => {
+  useEffect(() => {
+    onMount();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  return null;
+};