+++ front/app/utils/moduleUtils.ts
@@ -358,8 +358,17 @@ export type OutletsPropertyMap = {
   };
   'app.containers.LandingPage.EventsWidget': Record<string, any>;
   'app.containers.Admin.settings.customize.eventsSectionEnd': {
-    checked: boolean;
-    onChange: () => void;
+    getSetting: (settingName: string) => any;
+    setParentState: (state: any) => void;
+  };
+  'app.containers.Admin.settings.customize.Events': {
+    onMount: () => void;
+  };
+  'app.containers.Admin.settings.customize.AllInput': {
+    onMount: () => void;
+  };
+  'app.containers.Admin.initiatives.settings.EnableSwitch': {
+    onMount: () => void;
   };
   'app.containers.Admin.settings.customize.headerSectionStart': {
     latestAppConfigSettings:
@@ -547,6 +556,7 @@ export const insertConfiguration = <T extends { name: string }>({
   configuration,
   insertAfterName,
   insertBeforeName,
+  removeName,
 }: InsertConfigurationOptions<T>) => (items: T[]): T[] => {
   const itemAlreadyInserted = items.some(
     (item) => item.name === configuration.name
@@ -568,9 +578,19 @@ export const insertConfiguration = <T extends { name: string }>({
     items.splice(insertIndex, 1);
   }
 
-  return [
+  const newItems = [
     ...items.slice(0, insertIndex),
     configuration,
     ...items.slice(insertIndex),
   ];
+
+  if (removeName) {
+    const removeIndex = newItems.findIndex((item) => removeName === item.name);
+
+    if (removeIndex > -1) {
+      newItems.splice(removeIndex, 1);
+    }
+  }
+
+  return newItems;
 };